Unify mesh input
This modifies the MESH namelist to handle import of both the main
mesh and the EM mesh. There is a new `mesh_type` variable; the
default is 'unstructured' which is the type of the main mesh. The
EM mesh is now identified using the 'simplicial' (equivalently
'electromagnetics') mesh type.  The change also supports input of
arbitrarily many meshes, though additional meshes (of either type)
must be named. This is currently unused, but will support future
capabilities of the physics driver.

As part of harmonizing the mesh input, support for obsolete "gap"
element blocks has been removed, and the `coordinate_scale_factor`
variable is renamed to `coord_scale_factor`.