Import block, node and side set names from Exodus
This imports the element block, side set, and node set names
from the exodus mesh file and adds them to the unstructured
and simplicial mesh data structures. Where these are not
defined, an empty (0-length) name is assigned to the mesh
structures.

For internally generated meshes all names are empty.

No use is currently made of the names.